+++ ChangeLog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - added function 'io/slurp-stream'
 - added function 'io/spit-stream'
+- added function 'io/delete-file-on-exit'
 - added try resources block 'try-with'
 
 ### fixed